Added documentation to all classes except for Survey.java. Moved Answer.java to the AccessingDataJpaApplication folder and made Answer.java an Entity. Added some methods to Answer.java. Removed unnecessary Bean method from AccessingDataJpaApplication.java. Tested to make sure the program still works (it does).
