+++ src/main/java/minisurveymonkey/accessingdatajpa/QuestionRepository.java
@@ -8,7 +8,13 @@ import java.util.List;
 @Service
 public interface QuestionRepository extends CrudRepository<Question, Long> {
 
+    /* Searches for the question String, returns
+     * questionType, options, the ID, and the question itself.
+     */
     List<Question> findByQuestion(String question);
 
+    /* Searches for the ID of a question, returns
+     * questionType, options, the ID, and the question itself.
+     */
     Question findById(long id);
 }
