+++ src/main/java/minisurveymonkey/accessingdatajpa/QuestionService.java
@@ -3,7 +3,13 @@ package minisurveymonkey.accessingdatajpa;
 import java.util.List;
 
 public interface QuestionService {
+    /*
+     * Lists all questions saved to the repository.
+     */
     List<Question> findAll();
 
-    void save(Question buddy);
+    /*
+     * Saves a question to the repository.
+     */
+    void save(Question question);
 }
