+++ src/main/java/minisurveymonkey/accessingdatajpa/QuestionServiceImplement.java
@@ -8,16 +8,19 @@ import java.util.List;
 @Service
 public class QuestionServiceImplement implements QuestionService {
 
+    // Creates repository so we can save questions there.
     @Autowired
     private QuestionRepository questionRepository;
 
+    // Finds all questions saved to repository.
     @Override
     public List<Question> findAll() {
         return (List<Question>) questionRepository.findAll();
     }
 
+    // Save question to the repository.
     @Override
-    public void save(Question buddy) {
-        questionRepository.save(buddy);
+    public void save(Question question) {
+        questionRepository.save(question);
     }
 }
