+++ src/main/resources/templates/public/process.js
@@ -66,6 +66,7 @@ function createStuff() {
                 */
 }
 
+// Creates entries for the user to enter what options they want.
 function askAgain(num) {
     for (let i = 0; i < num; i++) {
         var p = document.createElement('p');
@@ -78,18 +79,28 @@ function askAgain(num) {
     }
 }
 
+// SETS THE NUMBER OF QUESTIONS.
 function setNum(num) {
     numberOfQuestions = num;
 }
 
+// HIDES THE "OPTIONS" SECTION IF THE USER CHOOSES A TEXT QUESTION.
 function hide() {
     document.getElementById('options').style.display = "none";
 }
 
+/* UNHIDES THE "OPTIONS" SECTION IF USER CHOOSES A RADIO BUTTONS QUESTION
+ * OR CHECKBOX QUESTION.
+ */
 function unHide() {
     document.getElementById('options').style.display = "block";
 }
 
+/* SETS THE NUMBER OF OPTIONS TO CHOOSE FROM.
+ * EXAMPLE... 3 OPTIONS IN A "FAVOURITE FOOD" RADIO BUTTON QUESTION: PIZZA,
+ * BUTTER CHICKEN, OR FAJITAS.
+ * PRINTS OUT A NUMBER OF OPTIONS FOR THE USER TO CUSTOMISE.
+ */
 function setOptionNum(val) {
     var num = parseInt(val);
     console.log(num);
@@ -106,6 +117,7 @@ function setOptionNum(val) {
     }
 }
 
+// RETRIEVES THE DATA FROM THE FORM AND OPENS A NEW PAGE
 function postData() {
     var val = document.getElementById("form").value;
